+++ src/hotmap/demo.ts
@@ -1,10 +1,13 @@
 import * as d3 from 'd3';
-import { Heatmap } from './heatmap';
+import { Heatmap, formatDataItem } from './heatmap';
 
 
 export function demo(divElementOrId: HTMLDivElement | string) {
-    // const hm = Heatmap.create();  // Heatmap<number, number, number>
+    const hm = Heatmap.create();  // Heatmap<number, number, number>
+    hm.render(divElementOrId);
+}
 
+export function demo2(divElementOrId: HTMLDivElement | string) {
     const items = [
         { col: 1, row: 'A', score: 0.0 },
         { col: 1, row: 'B', score: 0.2 },
@@ -32,6 +35,7 @@ export function demo(divElementOrId: HTMLDivElement | string) {
 
     const colorScale = d3.scaleLinear([0, 0.5, 1], ['#eeeeee', 'gold', 'red']);
     hm.setColor(d => colorScale(d.score));
+    hm.setTooltip((d, x, y, xIndex, yIndex) => `<div style="font-weight: bold; margin-bottom: 0.5em;">${formatDataItem(d)}</div>Column ${x}, Row ${y}<br>Indices [${xIndex},${yIndex}]`);
 
     hm.render(divElementOrId);
 }