+++ src/hotmap/demo.ts
@@ -37,20 +37,20 @@ export function demo2(divElementOrId: HTMLDivElement | string) {
     hm.setColor(d => colorScale(d.score));
     hm.setTooltip((d, x, y, xIndex, yIndex) => `<div style="font-weight: bold; margin-bottom: 0.5em;">${formatDataItem(d)}</div>Column ${x}, Row ${y}<br>Indices [${xIndex},${yIndex}]`);
     hm.setFilter((d, x, y, xIndex, yIndex) => d.score > 0);
-    hm.render(divElementOrId);
     setTimeout(() => hm.setFilter(undefined), 2000);
+    hm.setVisualParams({ xGapPixels: 0, yGapPixels: 0 });
     hm.events.hover.subscribe(e => {
         if (!e) {
             console.log('hovering nothing');
         } else {
-            console.log('hovering', e.datum, e.x, e.y, e.xIndex, e.yIndex);
+            console.log('hovering', e.datum, e.x, e.y, e.xIndex, e.yIndex, e.sourceEvent);
         }
     });
     hm.events.click.subscribe(e => {
         if (!e) {
             console.log('clicking nothing');
         } else {
-            console.log('clicking', e.datum, e.x, e.y, e.xIndex, e.yIndex);
+            console.log('clicking', e.datum, e.x, e.y, e.xIndex, e.yIndex, e.sourceEvent);
         }
     });
     hm.events.zoom.subscribe(e => {
@@ -60,5 +60,6 @@ export function demo2(divElementOrId: HTMLDivElement | string) {
             console.log('zooming', e);
         }
     });
+    hm.render(divElementOrId);
 }