+++ src/hotmap/utils.ts
@@ -1,4 +1,4 @@
-import { isEqual } from 'lodash';
+import { isEqual, isNil } from 'lodash';
 import { BehaviorSubject } from 'rxjs';
 import { BoxSize } from './scales';
 
@@ -27,8 +27,15 @@ export function kebabCase(str: string) {
     return str.replace(/([A-Z])/g, '-$1').toLowerCase();
 }
 
-export function nextIfChanged<T>(subject: BehaviorSubject<T>, newValue: T) {
-    if (!isEqual(subject.getValue(), newValue)) {
+export function nextIfChanged<T>(subject: BehaviorSubject<T>, newValue: T, key: ((value: T) => any) = (v => v)) {
+    if (!isEqual(key(subject.getValue()), key(newValue))) {
         subject.next(newValue);
     }
+}
+
+/** Return the smallest of the given numbers. Ignore any nullish values. If all values are nullish, return `undefined`. */
+export function minimum(...values: (number | null | undefined)[]): number | undefined {
+    const definedValues = values.filter(x => !isNil(x)) as number[];
+    if (definedValues.length > 0) return Math.min(...definedValues);
+    else return undefined;
 }